pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = "Client"
authors = ["Georgios Adzhygai"]
license = "MIT"
packages = [
    { include = "client", from = "src" },
]

[tool.poetry.dependencies]
python = "3.11.7"
django-environ = "^0.11.2"
pandas = "^2.1.4"
pyarrow = "^14.0.2"
fastparquet = "^2023.10.1"

[tool.poetry.group.service]
optional = true

[tool.poetry.group.service.dependencies]
scipy = "^1.11.1"
urllib3 = "^2.0.3"
redis = "^5.0.1"
uvicorn = "^0.25.0"
pydantic = "^2.5.3"
fastapi = "^0.108.0"
django-environ = "^0.11.2"
geopandas = "^0.14.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^23.12.1"
pytest = "^7.4.3"
pylint = "^3.0.3"

[tool.isort]
profile = "black"

[tool.black]
line-length = 80
target-version = ["py311"]

[tool.pylint]
max-line-length = 80
disable = ["missing-docstring", "too-few-public-methods", "redefined-outer-name", "fixme"]
source-roots = ["src"]

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
